requirements.txt
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
Deprecated==1.2.13
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
pycparser==2.21
PyJWT==2.5.0
PyNaCl==1.5.0
requests==2.28.1
smmap==5.0.0
urllib3==1.26.12
wrapt==1.14.1
